+++ src/components/AssetList/AssetListItem.tsx
@@ -2,6 +2,7 @@ import { observer } from 'mobx-react-lite';
 import { ListItem, ListItemIcon, ListItemText, ListItemProps } from '@cere-wallet/ui';
 import { Asset } from '~/stores';
 import { CoinIcon } from '../CoinIcon';
+import { useBalanceStore } from '~/hooks';
 
 export type AssetListItemProps = ListItemProps & {
   asset: Asset;
@@ -10,6 +11,8 @@ export type AssetListItemProps = ListItemProps & {
 export const AssetListItem = ({ asset, ...props }: AssetListItemProps) => {
   const { ticker, displayName, network, balance } = asset;
 
+  const { getUsdBalance } = useBalanceStore();
+
   return (
     <ListItem {...props}>
       <ListItemIcon inset>
@@ -17,7 +20,13 @@ export const AssetListItem = ({ asset, ...props }: AssetListItemProps) => {
       </ListItemIcon>
 
       <ListItemText primary={displayName} secondary={network} />
-      {balance !== undefined && <ListItemText align="right" primary={+balance.toFixed(2)} secondary="- USD" />}
+      {balance !== undefined && (
+        <ListItemText
+          align="right"
+          primary={+balance.toFixed(2)}
+          secondary={`$${getUsdBalance(ticker, balance).toFixed(2)} USD`}
+        />
+      )}
     </ListItem>
   );
 };